+++ src/index.less
@@ -5,6 +5,9 @@ body {
   height: 100%;
   font-size: 14px;
   background: #faf9ff;
+  a {
+    text-decoration: none;
+  }
 }
 .no_words_tips {
   position: absolute;
